refactor paddleaudio, test=doc
