+++ paddleaudio/CHANGELOG.md
@@ -1,1 +1,4 @@
 # Changelog
+
+Date: 2022-2-25, Author: Hui Zhang.
+  - Refactor architecture.
