+++ paddleaudio/paddleaudio/io/__init__.py
@@ -1,0 +1,6 @@
+from .audio import save_wav
+from .audio import load
+from .audio import normalize
+from .audio import to_mono
+from .audio import resample
+from .audio import depth_convert
