+++ paddleaudio/paddleaudio/utils/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
 #
-# Licensed under the Apache License, Version 2.0 (the "License");
+# Licensed under the Apache License, Version 2.0 (the "License"
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
@@ -11,4 +11,20 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .audio import *
+
+from .env import USER_HOME
+from .env import PPAUDIO_HOME
+from .env import MODEL_HOME
+from .env import DATA_HOME
+
+from .download import decompress
+from .download import download_and_decompress
+from .download import load_state_dict_from_url
+
+from .error import ParameterError
+
+from .log import logger
+from .log import Logger
+
+from .time import Timer
+from .time import seconds_to_hms
