+++ paddleaudio/paddleaudio/utils/time.py
@@ -14,6 +14,10 @@
 import math
 import time
 
+__all__ = [
+    'Timer',
+    'seconds_to_hms',
+]
 
 class Timer(object):
     '''Calculate runing speed and estimated time of arrival(ETA)'''
